refactor: 模块化overall界面preview区域
